database added, creating, updating, deleting and completing a task

+++ src/routes.js
@@ -1,65 +1,114 @@
-import { randomUUID } from "node:crypto";
-import { Database } from "./database.js";
-import { buildRoutePath } from "./utils/build-route-path.js";
+import { randomUUID } from 'node:crypto';
+import { Database } from './database.js';
+import { buildRoutePath } from './utils/build-route-path.js';
 
 const database = new Database();
 
 export const routes = [
   {
-    method: "POST",
-    path: buildRoutePath("/tasks"),
+    method: 'POST',
+    path: buildRoutePath('/tasks'),
     handler: (req, res) => {
       const { title, description } = req.body;
+
+      dateNow = Date.now();
 
       const task = {
         id: randomUUID(),
         title,
         description,
+        completed_at: null,
+        created_at: dateNow,
+        updated_at: dateNow,
       };
 
-      database.insert("tasks", task);
+      database.insert('tasks', task);
 
       return res.writeHead(201).end();
     },
   },
   {
-    method: "GET",
-    path: buildRoutePath("/tasks"),
+    method: 'GET',
+    path: buildRoutePath('/tasks'),
     handler: (req, res) => {
-      const tasks = database.select("tasks");
+      const tasks = database.select('tasks');
       return res.end(JSON.stringify(tasks));
     },
   },
 
   {
-    method: "PUT",
-    path: buildRoutePath("/tasks/:id"),
+    method: 'PUT',
+    path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
       const { id } = req.params;
       const { title, description } = req.body;
 
-      database.update("tasks", id, { title, description });
-      return res.writeHead(204).end();
+      const tasks = database.select('tasks');
+
+      const rowIndex = tasks.findIndex((row) => row.id === id);
+
+      if (rowIndex > -1) {
+        taskToUpdate = tasks[rowIndex];
+
+        dateNow = Date.now();
+
+        taskToUpdate = {
+          ...taskToUpdate,
+          title,
+          description,
+          updated_at: dateNow,
+        };
+
+        database.update('tasks', id, taskToUpdate);
+        return res.writeHead(204).end();
+      } else {
+        return res.writeHead(404).end();
+      }
     },
   },
   {
-    method: "DELETE",
-    path: buildRoutePath("/tasks/:id"),
+    method: 'DELETE',
+    path: buildRoutePath('/tasks/:id'),
     handler: (req, res) => {
       const { id } = req.params;
 
-      database.delete("tasks", id);
-      return res.writeHead(204).end();
+      const tasks = database.select('tasks');
+
+      const rowIndex = tasks.findIndex((row) => row.id === id);
+
+      if (rowIndex > -1) {
+        database.delete('tasks', id);
+        return res.writeHead(204).end();
+      } else {
+        return res.writeHead(404).end();
+      }
     },
   },
   {
-    method: "PATCH",
-    path: buildRoutePath("/tasks/:id/complete"),
+    method: 'PATCH',
+    path: buildRoutePath('/tasks/:id/complete'),
     handler: (req, res) => {
       const { id } = req.params;
 
-      database.complete("tasks", id);
-      return res.writeHead(204).end();
+      const tasks = database.select('tasks');
+
+      const rowIndex = tasks.findIndex((row) => row.id === id);
+
+      if (rowIndex > -1) {
+        taskToUpdate = tasks[rowIndex];
+
+        dateNow = Date.now();
+
+        taskToUpdate = {
+          ...taskToUpdate,
+          completed_at: dateNow,
+        };
+
+        database.update('tasks', id, taskToUpdate);
+        return res.writeHead(204).end();
+      } else {
+        return res.writeHead(404).end();
+      }
     },
   },
 ];
